+++ backend/controllers/ProjectController.php
@@ -4,6 +4,7 @@ namespace backend\controllers;
 
 use common\models\Project;
 use common\models\Users;
+use common\models\History;
 use yii\web\Controller;
 use Yii;
 use yii\web\ForbiddenHttpException;
@@ -34,27 +35,26 @@ class ProjectController extends Controller
     }
 
     public function actionAddProject()
-    {
-        if (Yii::$app->user->can('create')) {
+ {
+        if ( Yii::$app->user->can( 'create' ) ) {
             $model = new Project();
-    
-            if ($model->load($this->request->post())) {
+
+            if ( $model->load( $this->request->post() ) ) {
                 // Combine start_date and end_date into a formatted duration string
                 $model->duration = $model->start_date . ' to ' . $model->end_date;
-    
-                if ($model->save()) {
-                    return $this->redirect(['view-project', 'project_id' => $model->project_id]);
+
+                if ( $model->save() ) {
+                    return $this->redirect( [ 'view-project', 'project_id' => $model->project_id ] );
                 }
             }
-    
-            return $this->render('add-project', [
+
+            return $this->render( 'add-project', [
                 'model' => $model,
-            ]);
+            ] );
         } else {
             throw new ForbiddenHttpException;
         }
     }
-    
 
     public function actionUpdate( $project_id )
  {
@@ -76,9 +76,18 @@ class ProjectController extends Controller
     public function actionDelete( $project_id )
  {
         if ( Yii::$app->user->can( 'delete' ) ) {
-            $this->findModel( $project_id )->delete();
+            // Check if there are associated giz_interventions_history records
+            $hasInterventionsHistory = History::find()->where( [ 'project_id' => $project_id ] )->exists();
 
-            return $this->redirect( [ '/site/index' ] );
+            if ( $hasInterventionsHistory ) {
+                Yii::$app->session->setFlash( 'error', '--------------------------------------Cannot delete the `Project` because there are associated records in `Intervention History`-------------------------' );
+
+            } else {
+                $this->findModel( $project_id )->delete();
+            }
+
+            return $this->redirect( [ 'view-project' ] );
+
         } else {
             throw new ForbiddenHttpException;
         }
